+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV7CRSERT08234DB1760N2D3"
+        "01KV7W75A232WFR3EN22KJCVE6"
     )
     r"""The unique identifier of the data source"""