+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV7CRSENPNC9E00JXBX1ACPN"
+        "01KV7W75A0C20B39DZQ1Z7TR7P"
     )
     r"""The unique identifier of the data source"""