+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV7CRSEQWNK6K71F5S8XN8T4"
+        "01KV7W75A19DGFA62K24WNMBV1"
     )
     r"""The unique identifier of the data source"""