+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV7CRSEVP0Q8ZAFJQ2JG2GMP"
+        "01KV7W75A5FZ7BV4K07JBBPMJF"
     )
     r"""The unique identifier of the data source"""