+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KV7CRSB0QHYKS9Q3XA47TBYG"
+    id: Optional[str] = "01KV7W756XDC1FAHC188TRNZ55"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KV7CRSARJNEP8M5AF3RYJVX0"
+        "tool_01KV7W756QSAFQZVCWCS9G03CM"
     )
 
     display_name: Optional[str] = None
@@ -1488,7 +1488,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KV7CRSAQHV87782DBQKPY8NY"
+    id: Optional[str] = "01KV7W756P4Y4K4FDN4V9S8H19"
 
     description: Optional[str] = None
 
@@ -1632,7 +1632,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KV7CRSAPEQV1F2B5VVWNK358"
+        "tool_01KV7W756N3TNKV2XMRBK3VNJY"
     )
 
     display_name: Optional[str] = None
@@ -1932,7 +1932,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KV7CRSAM3NZ2W5BF27S5ZN85"
+        "tool_01KV7W756KF9K34NG4BDJFPFPD"
     )
 
     display_name: Optional[str] = None
@@ -2126,7 +2126,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KV7CRSAJ6BN6CK93EGQ54KJ3"
+        "tool_01KV7W756H2DSJCH7NWMMJ9MV8"
     )
 
     display_name: Optional[str] = None
@@ -2324,7 +2324,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KV7CRSAGPMCD5TB8R5JFKSDH"
+        "tool_01KV7W756FVKDT6TARB1ZGEVFT"
     )
 
     display_name: Optional[str] = None